+++ requirements.txt
@@ -1,5 +1,6 @@
 aiohttp
 PyNaCl
+youtube_dl
 python-tds
 sqlalchemy-pytds
 sqlalchemy
