+++ src/framework/message/__init__.py
@@ -1,0 +1,7 @@
+"""~ sub-package ~
+@Info:
+    This sub-package contains the message definitions.
+    It is a seperate package to reduce number of lines per file."""
+from .base import *
+from .text_based import *
+from .voice_based import *
