+++ src/framework/message/text_based.py
@@ -1,416 +1,23 @@
-"""
-    ~   message    ~
-    This module contains the definitions regarding the xxxMESSAGE class and
-    all the functionality for sending data into discord channels.
-"""
-from    typing import List, Union, Literal, Iterable, Any
-from    .dtypes import *
-from    .tracing import *
-from    .const import *
-from    .timing import *
-from    . import client
-from    . import sql
-import  random
-import  asyncio
-import  _discord as discord
+"""~  text message related ~
+@Info:
+    Contains definitions for message classes that are text based (TextMESSAGE & DirectMESSAGE)."""
+
+from   .base      import *
+from   ..         import client
+from   ..         import sql
+from   ..dtypes   import *
+from   ..tracing  import *
+from   ..const    import *
+from   ..         import core
+from   typing     import List, Iterable, Union, Literal
+import asyncio
+import _discord as discord
 
 
 __all__ = (
     "TextMESSAGE",
-    "VoiceMESSAGE",
     "DirectMESSAGE"
 )
-
-
-class GLOBALS:
-    """ ~  GLOBALS  ~
-        @Info: Contains the globally needed variables"""
-    is_user = True
-
-
-# Dummy classes for message mods, to use with sql.register_type decorator
-@sql.register_type("MessageMODE")
-class _:
-    __logname__ = "send"
-@sql.register_type("MessageMODE")
-class _:
-    __logname__ = "edit"
-@sql.register_type("MessageMODE")
-class _:
-    __logname__ = "clear-send"
-
-
-class BaseMESSAGE:
-    """~  BaseMESSAGE  ~
-        @Info:
-        This is the base class for all the different classes that
-        represent a message you want to be sent into discord."""
-
-    __slots__ = (
-        "initialized",
-        "randomized_time",
-        "period",
-        "random_range",
-        "timer",
-        "force_retry",
-        "data"
-    )
-
-    # The "__valid_data_types__" should be implemented in the INHERITED classes.
-    # The set contains all the data types that the class is allowed to accept, this variable
-    # is then checked for allowed data types in the "initialize" function bellow.
-    __valid_data_types__ = {}
-
-    def __init__(self,
-                start_period : Union[float,None],
-                end_period : float,
-                data,
-                start_now : bool=True):
-        # If start_period is none -> period will not be randomized
-        if start_period is None:            
-            self.randomized_time = False
-            self.period = end_period
-        else:
-            self.randomized_time = True
-            self.random_range = (start_period, end_period)
-            self.period = random.randrange(*self.random_range)
-
-        self.timer = TIMER()
-        self.force_retry = {"ENABLED" : start_now, "TIME" : 0}
-        self.data = data
-        self.initialized = False
-
-    def generate_exception(self, 
-                           status: int,
-                           code: int,
-                           description: str,
-                           cls: discord.HTTPException):
-        """ ~ method ~
-        @Name: generate_exception
-        @Info: Generates a discord.HTTPException inherited class exception object
-        @Param:
-            status: int ~ Atatus code of the exception
-            code: int ~ Actual error code
-            description: str ~ The textual description of the error
-            cls: discord.HTTPException ~ Inherited class to make exception from"""
-        resp = Exception()
-        resp.status = status
-        resp.status_code = status
-        resp.reason = cls.__name__
-        ex = cls(resp, {"message" : description, "code" : code})
-        return ex
-
-    def generate_log_context(self):
-        """ ~ method ~
-        @Name: generate_log_context
-        @Info:
-            This method is used for generating a dictionary (later converted to json) of the
-            data that is to be included in the message log. This is to be implemented inside the
-            inherited classes."""
-        raise NotImplementedError
-    
-    def get_data(self) -> dict:
-        """ ~ method ~
-        @Name:  get_data
-        @Info: Returns a dictionary of keyword arguments that is then expanded
-               into other functions (send_channel, generate_log)
-               This is to be implemented in inherited classes due to different data_types"""
-        raise NotImplementedError
-
-    def is_ready(self) -> bool:
-        """ ~ method ~
-        @Name:   is_ready
-        @Param:  void
-        @Info:   This method returns bool indicating if message is ready to be sent"""
-        return (not self.force_retry["ENABLED"] and self.timer.elapsed() > self.period or
-                self.force_retry["ENABLED"] and self.timer.elapsed() > self.force_retry["TIME"])
-
-    def reset_timer(self) -> None:
-        """ ~ method ~
-        @Name: restart_time
-        @Info: Resets internal timer (and force period)"""
-        self.timer.reset()
-        self.timer.start()
-        self.force_retry["ENABLED"] = False
-        if self.randomized_time is True:
-            self.period = random.randrange(*self.random_range)
-
-    async def send_channel(self) -> dict:
-        """                  ~ send_channel ~
-            @Info:
-            Sends data to a specific channel, this is seperate from send
-            for eaiser implementation of simmilar inherited classes
-            @Return:
-            The method returns a dictionary containing : {"success": bool, "reason": discord.HTTPException}
-            """
-        raise NotImplementedError
-
-    async def send(self) -> dict:
-        """ ~ async method ~
-        Name:   send to channels
-        Param:  void
-        Info:   This function should be implemented in the inherited class
-                and should send the message to all the channels."""
-        raise NotImplementedError
-
-    async def initialize_channels(self) -> bool:
-        """ ~ async method ~
-        @Name: initialize_channels
-        @Info: This method initializes the implementation specific
-               api objects and checks for the correct channel inpit context."""
-        raise NotImplementedError
-
-    async def initialize_data(self) -> bool:
-        """ ~ async method ~
-        @Name:  initialize_data
-        @Info:  This method checks for the correct data input to the xxxMESSAGE
-                object. The expected datatypes for specific implementation is
-                defined thru the static variable __valid_data_types__"""
-        # Check for correct data types of the MESSAGE.data parameter
-        if not isinstance(self.data, FunctionBaseCLASS):
-            # This is meant only as a pre-check if the parameters are correct so you wouldn't eg. start
-            # sending this message 6 hours later and only then realize the parameters were incorrect.
-            # The parameters also get checked/parsed each period right before the send.
-
-            # Convert any arguments passed into a list of arguments
-            if isinstance(self.data, (list, tuple, set)):
-                self.data = list(self.data)   # Convert into a regular list to allow removal of items
-            else:
-                self.data = [self.data]       # Place into a list for iteration, to avoid additional code
-
-            # Check all the arguments
-            for data in self.data[:]:
-                # Check all the data types of all the passed to the data parameter.
-                # If class does not match the allowed types, then the object is removed.
-                # The for loop iterates thru a shallow copy (sliced list) of data_params to allow removal of items
-                # without affecting the iteration (would skip elements without a copy or use of while loop).
-
-                # The inherited classes MUST DEFINE THE "__valid_data_types__" inside the class which should be a set of the allowed data types
-
-                if (
-                        type(data) not in type(self).__valid_data_types__
-                    ):
-                    if isinstance(self.data, FunctionBaseCLASS):
-                        trace(f"The function can only be used on the data parameter directly, not in a list\nFunction: {data.func_name}", TraceLEVELS.ERROR)
-                        self.data.clear()
-                        break
-                    else:
-                        trace(f"INVALID DATA PARAMETER PASSED!\nArgument is of type : {type(data).__name__}\nSee README.md for allowed data types", TraceLEVELS.WARNING)
-                        self.data.remove(data)
-
-            return len(self.data) > 0
-
-        return True
-
-    async def initialize(self, **options) -> bool:
-        """ ~ async method ~
-        @Name: initialize
-        @Info:
-            The initialize method initilizes the message object.
-        @Params:
-            options :: custom keyword arguments, this differes from inher. to inher. class that
-                       is inherited from the BaseGUILD class and must be matched in the inherited class from BaseMESSAGE that
-                       you want to use in that specific inherited class from BaseGUILD class"""
-        if self.initialized:
-            return True
-
-        if not await self.initialize_channels(**options):
-            return False
-
-        if not await self.initialize_data():
-            return False
-
-        self.initialized = True
-        return True
-
-@sql.register_type("MessageTYPE")
-class VoiceMESSAGE(BaseMESSAGE):
-    """ ~ BaseMESSAGE class ~
-    Name: VoiceMESSAGE
-    Info: The VoiceMESSAGE object containts parameters which describe behaviour and data that will be sent to the channels.
-    Params:
-        Start Period , End Period (start_period, end_period) :: These 2 parameters specify the period on which the messages will be played:
-            Start Period can be either:
-                :: None :: Messages will be sent on intervals specified by End period,
-                :: Integer >= 0 :: Messages will be sent on intervals randomly chosen between Start period and End period,
-                                   where the randomly chosen intervals will be re::randomized after each sent message.
-        Data (data) :: The data parameter is the actual data that will be sent using discord's API. The data types of this parameter can be:
-                        - Path to an audio file (str)
-                        - Function that accepts any amount of parameters and returns any of the above types.
-                          To pass a function, YOU MUST USE THE framework.data_function decorator on the function before
-                          passing the function to the framework.
-        Channel IDs (channel_ids) :: List of IDs of all the channels you want data to be sent into.
-        Start Now (start_now) :: A bool variable that can be either True or False. If True, then the framework will send the message
-                                 as soon as it is run and then wait it's period before trying again. If False, then the message will
-                                 not be sent immediatly after framework is ready, but will instead wait for the period to elapse."""
-
-    __slots__ = (
-        "randomized_time",
-        "period",
-        "random_range",
-        "data",
-        "channels",
-        "timer",
-        "force_retry",
-    )
-
-    __logname__ = "VoiceMESSAGE"
-    __valid_data_types__ = {AUDIO}  # This is used in the BaseMESSAGE.initialize() to check if the passed data parameters are of correct type
-
-    def __init__(self, start_period: Union[float, None],
-                 end_period: float,
-                 data: AUDIO,
-                 channel_ids: Iterable[int],
-                 start_now: bool = True):
-
-        super().__init__(start_period, end_period, start_now)
-        self.data = data
-        self.channels = list(set(channel_ids)) # Auto remove duplicates
-
-    def generate_log_context(self,
-                             audio: AUDIO,
-                             succeeded_ch: List[discord.VoiceChannel],
-                             failed_ch: List[dict]):
-        """ ~ method ~
-        @Name: generate_log_context
-        @Param:
-            audio: AUDIO -- The audio that was streamed to the channels
-            succeeded_ch: List[discord.VoiceChannel] -- list of the successfuly streamed channels,
-            failed_ch: List[dict]   -- list of dictionaries contained the failed channel and the Exception
-        @Info:
-            Generates a dictionary containing data that will be saved in the message log"""
-
-        succeeded_ch = [{"name": str(channel), "id" : channel.id} for channel in succeeded_ch]
-        failed_ch = [{"name": str(entry["channel"]), "id" : entry["channel"].id,
-                     "reason": str(entry["reason"])} for entry in failed_ch]
-        return {
-            "sent_data": {
-                "streamed_audio" : audio.filename
-            },
-            "channels": {
-                "successful" : succeeded_ch,
-                "failed": failed_ch
-            },
-            "type" : type(self).__name__
-        }
-
-    def get_data(self) -> dict:
-        """ ~ method ~
-        @Name:  get_data
-        @Info: Returns a dictionary of keyword arguments that is then expanded
-               into other functions (send_channel, generate_log)"""
-        data = None
-        _data_to_send = {}
-        data = self.data.get_data() if isinstance(self.data, FunctionBaseCLASS) else self.data
-        if data is not None:
-            if not isinstance(data, (list, tuple, set)):
-                data = (data,)
-            for element in data:
-                if isinstance(element, AUDIO):
-                    _data_to_send["audio"] = element
-        return _data_to_send
-
-    async def initialize_channels(self) -> bool:
-        """ ~ async method ~
-        @Name: initialize_channels
-        @Info: This method initializes the implementation specific
-               api objects and checks for the correct channel inpit context.
-        @Return: Bool - True on success"""
-        ch_i = 0
-        cl = client.get_client()
-        while ch_i < len(self.channels):
-            channel_id = self.channels[ch_i]
-            channel = cl.get_channel(channel_id)
-            self.channels[ch_i] = channel
-
-            if channel is None:
-                trace(f"Unable to get channel from ID {channel_id}", TraceLEVELS.ERROR)
-                self.channels.remove(channel)
-            elif type(channel) is not discord.VoiceChannel:
-                trace(f"VoiceMESSAGE object got ID ({channel_id}) for {type(channel).__name__}, but was expecting {discord.VoiceChannel.__name__}", TraceLEVELS.WARNING)
-                self.channels.remove(channel)
-            else:
-                ch_i += 1
-
-        return len(self.channels) > 0
-
-    async def send_channel(self,
-                           channel: discord.VoiceChannel,
-                           audio: AUDIO) -> dict:
-        """ ~ async method ~
-        @Name : send_channel
-        @Info:
-            Streams audio to specific channel
-        @Return:
-        - dict:
-            - "success" : bool ~ True if successful, else False
-            - "reason"  : Exception ~ Only present if "success" is False,
-                            contains the Exception returned by the send attempt."""
-        stream = None
-        voice_client = None
-        try:
-            # Check if client has permissions before attempting to join
-            ch_perms = channel.permissions_for(channel.guild.get_member(client.get_client().user.id))
-            if not all([ch_perms.connect, ch_perms.stream, ch_perms.speak]):
-                raise self.generate_exception(403, 50013, "You lack permissions to perform that action", discord.Forbidden)
-
-            # Try to open file first as FFMpegOpusAudio doesn't raise exception if file does not exist
-            with open(audio.filename, "rb"):
-                pass
-            stream = discord.FFmpegOpusAudio(audio.filename)
-
-            voice_client = await channel.connect(timeout=C_VC_CONNECT_TIMEOUT)
-            voice_client.play(stream)
-            while voice_client.is_playing():
-                await asyncio.sleep(1)
-            return {"success": True}
-        except Exception as ex:
-            if client.get_client().get_channel(channel.id) is None:
-                ex = self.generate_exception(404, 10003, "Channel was deleted", discord.NotFound)
-            else:
-                ex = self.generate_exception(500, 0, "Timeout error", discord.HTTPException)
-            return {"success": False, "reason": ex}
-        finally:
-            if stream is not None:
-                stream.cleanup()
-            if voice_client is not None:
-                # Note (TODO): Should remove this in the future. Currently it disconnects instead moving to a different channel, because using
-                #         .move_to(channel) method causes some sorts of "thread leak" (new threads keep getting created, waiting for pycord to fix this).
-                await voice_client.disconnect()
-
-    async def send(self) -> Union[dict,  None]:
-        """" ~ async method ~
-        @Name send
-        @Info:
-            Streams audio into the chanels
-        @Return:
-            Returns a dictionary generated by the generate_log_context method
-            or the None object if message wasn't ready to be sent"""
-        _data_to_send = self.get_data()
-        if any(_data_to_send.values()):
-            errored_channels = []
-            succeded_channels= []
-
-            for channel in self.channels:
-                context = await self.send_channel(channel, **_data_to_send)
-                if context["success"]:
-                    succeded_channels.append(channel)
-                else:
-                    errored_channels.append({"channel":channel, "reason": context["reason"]})
-
-            # Remove any channels that returned with code status 404 (They no longer exist)
-            for data in errored_channels:
-                reason = data["reason"]
-                channel = data["channel"]
-                if isinstance(reason, discord.HTTPException):
-                    if (reason.status == 403 or
-                        reason.code in {10003, 50013} # Unknown, Permissions
-                    ):
-                        self.channels.remove(channel)
-                        trace(f"Channel {channel.name}(ID: {channel.id}) {'was deleted' if reason.code == 10003 else 'does not have permissions'}, removing it from the send list", TraceLEVELS.WARNING)
-
-            return self.generate_log_context(**_data_to_send, succeeded_ch=succeded_channels, failed_ch=errored_channels)
-        return None
 
 
 @sql.register_type("MessageTYPE")
@@ -635,7 +242,7 @@ class TextMESSAGE(BaseMESSAGE):
             # Send to channels
             for channel in self.channels:
                 # Clear previous messages sent to channel if mode is MODE_DELETE_SEND
-                if GLOBALS.is_user:
+                if core.GLOBALS.is_user:
                     await asyncio.sleep(RLIM_USER_WAIT_TIME)
                 context = await self.send_channel(channel, **data_to_send)
                 if context["success"]:
@@ -844,7 +451,7 @@ class DirectMESSAGE(BaseMESSAGE):
         # Parse data from the data parameter
         data_to_send = self.get_data()
         if any(data_to_send.values()):
-            if GLOBALS.is_user:
+            if core.GLOBALS.is_user:
                 await asyncio.sleep(RLIM_USER_WAIT_TIME)
             context = await self.send_channel(**data_to_send)
             if context["success"] is False:
@@ -858,14 +465,3 @@ class DirectMESSAGE(BaseMESSAGE):
             return self.generate_log_context(**data_to_send, **context)
 
         return None
-
-
-async def initialize(is_user: bool) -> bool:
-    """ ~ async function ~
-    @Name: initialize
-    @Info:
-        Initializes the module.
-    @Parameters:
-        - is_user: bool ~ True if the class is used for a user, else False"""
-    GLOBALS.is_user = is_user
-    return True
